chore: 스타일 관련 코드 파일로 분리
