+++ src/components/common/BottomMenu/index.tsx
@@ -1,16 +1,16 @@
-import styled from '@emotion/styled'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
 import Image from 'next/image'
-import { BOTTOM_MENU_INFO } from '@/constants'
+import { BOTTOM_MENU_INFO } from '@/constants/menu'
 import { BottomMenuInfoProps } from '@/types/styles'
-import Divider from '../Divider'
+import St from './style'
+import Divider from '@components/common/Divider'
 
 const BottomMenu = () => {
   const { pathname } = useRouter()
   // console.log(pathname)
   return (
-    <TabsWrapper>
+    <St.Root>
       {BOTTOM_MENU_INFO.map(
         (
           {
@@ -32,21 +32,8 @@ const BottomMenu = () => {
           </>
         ),
       )}
-    </TabsWrapper>
+    </St.Root>
   )
 }
 
 export default BottomMenu
-
-const TabsWrapper = styled.div`
-  /*position: fixed;*/
-  bottom: 0;
-  width: inherit;
-
-  background-color: var(--color-green-400);
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-evenly;
-  padding: 1.8rem 0;
-`
