+++ src/components/common/BottomMenu/style.ts
@@ -1,0 +1,17 @@
+import styled from '@emotion/styled'
+
+const Root = styled.div`
+  /*position: fixed;*/
+  bottom: 0;
+  width: inherit;
+
+  background-color: var(--color-green-400);
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-evenly;
+  padding: 1.8rem 0;
+`
+const St = { Root }
+
+export default St
