+++ src/components/common/Divider/style.ts
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled'
+
+const Root = styled.div`
+  height: 3.5rem;
+  width: 0.1rem;
+  background-color: var(--color-green-200);
+`
+
+const St = { Root }
+
+export default St
