+++ src/components/common/Header/style.ts
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled'
+
+const Root = styled.div`
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  padding-right: 3.2rem;
+`
+const St = { Root }
+export default St
