+++ src/components/common/Island/Content/style.ts
@@ -1,0 +1,18 @@
+import styled from '@emotion/styled'
+
+const Root = styled.div`
+  position: relative;
+  height: 78rem;
+  width: 43rem;
+`
+const Items = {
+  Root: styled.div`
+    position: relative;
+    top: 0;
+    left: 0;
+    height: inherit;
+  `,
+}
+const St = { Root, Items }
+
+export default St
