+++ src/components/common/Layout/Layout/style.ts
@@ -1,0 +1,12 @@
+import styled from '@emotion/styled'
+
+const Root = styled.div`
+  width: 43rem;
+  scrollbar-width: none;
+  min-height: 100dvh;
+  margin: auto;
+  background-color: var(--color-yellow-100);
+`
+const St = { Root }
+
+export default St
