+++ src/components/common/Logo/index.tsx
@@ -1,4 +1,4 @@
-import { LOGO_TYPE_INFO } from '@/constants'
+import { LOGO_TYPE_INFO } from '@/constants/styles'
 import Image from 'next/image'
 import St from './style'
 import { LogoProps } from '@/types/styles'
@@ -7,9 +7,9 @@ const Logo = ({ type, padding }: LogoProps) => {
   const { src, width, height, alt } = LOGO_TYPE_INFO[type]
 
   return (
-    <St.LogoWrapper padding={padding ?? 0}>
+    <St.Root padding={padding ?? 0}>
       <Image src={src} width={width} height={height} alt={alt} priority />
-    </St.LogoWrapper>
+    </St.Root>
   )
 }
