+++ src/components/common/Logo/style.ts
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled'
 
 const St = {
-  LogoWrapper: styled.div<{ padding: number }>`
+  Root: styled.div<{ padding: number }>`
     width: fit-content;
     height: fit-content;
     padding: ${({ padding }) => `${padding}rem`};
