+++ src/components/common/Modal/Game/index.tsx
@@ -1,12 +1,11 @@
-import { GAME_START_POINT_BTN } from '@/constants'
-import { STATE_MODAL_TYPE_OVERLAY } from '@/constants/modal/state'
-import { FONTS } from '@/styles/font'
-import styled from '@emotion/styled'
+import { GAME_START_POINT_BTN } from '@/constants/styles'
+import { STATE_MODAL_TYPE_OVERLAY } from '@/constants/modal'
+
 import { useRouter } from 'next/router'
 
 import { useRef } from 'react'
-import { PointBtn } from '../../Button'
-import PortalModal from '../PortalModal'
+import PointBtn from '@/components/common/Button'
+import PortalModal from '@/components/common/Modal/PortalModal'
 import Image from 'next/image'
 import {
   GameStatus,
@@ -14,14 +13,15 @@ import {
   PlayStateProps,
   StateModalProps,
 } from '@/types/game'
-import { IMAGE_ADDRESS } from '@/constants'
+import { IMAGE_ADDRESS } from '@/constants/styles'
 import { getDuration } from '@/lib'
 import { useKeyEscape, useTimer, useOutsideClick } from '@/hooks'
+import St from './style'
 
 const Start = () => {
   return (
     <PointBtn item={GAME_START_POINT_BTN}>
-      <Text>게임 시작</Text>
+      <St.Text>게임 시작</St.Text>
     </PointBtn>
   )
 }
@@ -73,43 +73,21 @@ const StateModal = ({
   if (!isOpen || !router.isReady) return null
   return (
     <PortalModal>
-      <ModalWrapper color={STATE_MODAL_TYPE_OVERLAY[router.pathname]}>
+      <St.Root color={STATE_MODAL_TYPE_OVERLAY[router.pathname]}>
         {/*<button onClick={handleClose} className="close-btn">
           Close
         </button>*/}
-        <Content ref={contentRef}>
+        <St.Content ref={contentRef}>
           {type == GameStatus.MATCHED && content && (
             <Matched content={content} />
           )}
           {type == GameStatus.START && <Start />}
           {type == GameStatus.CLEAR && <Clear />}
           {type == GameStatus.RESULT && content && <Result content={content} />}
-        </Content>
-      </ModalWrapper>
+        </St.Content>
+      </St.Root>
     </PortalModal>
   )
 }
 
 export default StateModal
-
-const ModalWrapper = styled.div<{ color: string }>`
-  position: fixed;
-  inset: 0; /* inset sets all 4 values (top right bottom left) much like how we set padding, margin etc., */
-  background-color: ${({ color }) => `${color}`};
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  transition: all 0.3s ease-in-out;
-  overflow: hidden;
-  z-index: 999;
-`
-const Content = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`
-const Text = styled.div`
-  ${FONTS.h1}
-  color:var(--color-orange-300);
-`
