+++ src/components/common/Triangle/index.tsx
@@ -1,4 +1,4 @@
-import { TRIANGLE_TYPE_INFO } from '@/constants'
+import { TRIANGLE_TYPE_INFO } from '@/constants/styles'
 import { TriangleProps } from '@/types/styles'
 import Image from 'next/image'
