+++ src/components/pages/Home/UserPointBtn.tsx
@@ -1,5 +1,5 @@
-import { PointBtn } from '@/components/common/Button'
-import { USER_POINT_BTN } from '@/constants/styles/pointBtn'
+import PointBtn from '@/components/common/Button'
+import { USER_POINT_BTN } from '@/constants/styles'
 import { FONTS } from '@styles/font'
 import styled from '@emotion/styled'
 import { useSession } from 'next-auth/react'
