+++ src/components/pages/Signup/form/index.tsx
@@ -1,4 +1,4 @@
-import { FormBtn } from '@/components/common/Button'
+import { FormBtn } from '@/components/common/Button/style'
 import { StateModal } from '@/components/common/Modal'
 import { useModalActions, useModalStore } from '@/store/modal'
 import { FONTS } from '@/styles/font'
@@ -9,7 +9,7 @@ import {
   SIGNUP_FORM,
   SIGNUP_INITIAL_VALUES,
   SIGNUP_VALIDATION,
-} from '@/constants'
+} from '@/constants/form'
 import useDiscordOAuth from './useDiscordOAuth'
 import { useEffect } from 'react'
