+++ src/components/pages/Visit/finding/index.tsx
@@ -3,7 +3,7 @@ import VisitForm from './form'
 import Header from './header'
 import Image from 'next/image'
 import styled from '@emotion/styled'
-import { IMAGE_ADDRESS } from '@/constants'
+import { IMAGE_ADDRESS } from '@/constants/styles'
 
 const VistFinding = () => {
   return (
