--- src/constants/index.ts
@@ -1,57 +1,0 @@
-import { BACKGROUND_TYPE_INFO } from './styles/background'
-import { colors, shadows, gradients } from './styles/colors'
-import LOCAL_STORAGE_KEY from './localStorageKey'
-import { LOGO_TYPE_INFO } from './styles/logo'
-import { TRIANGLE_TYPE_INFO } from './styles/triangle'
-import { IMAGE_ADDRESS } from '@/constants/styles/imageAddress'
-import { Bounce } from './styles/animations'
-import { withAuthList, withOutAuthList, FALLBACK_URL } from './routeUrl'
-import { BOTTOM_MENU_INFO } from './menu/bottom'
-import { STATE_MODAL_TYPE_INFO } from './modal/state'
-import {
-  USER_POINT_BTN,
-  MENU_POINT_BTN,
-  GAME_START_POINT_BTN,
-  EDIT_TAB_POINT_BTN,
-} from './styles/pointBtn'
-import {
-  SIGNUP_FORM,
-  SIGNUP_INITIAL_VALUES,
-  SIGNUP_VALIDATION,
-} from './form/signup'
-import {
-  LS_EMAIL,
-  LOGIN_INITIAL_VALUES,
-  LOGIN_VALIDATION,
-  LOGIN_FORM,
-} from './form/login'
-import { VISIT_VALIDATION, VISIT_INITIAL_VALUE } from './form/visit'
-export {
-  IMAGE_ADDRESS,
-  BACKGROUND_TYPE_INFO,
-  LOCAL_STORAGE_KEY,
-  LOGO_TYPE_INFO,
-  TRIANGLE_TYPE_INFO,
-  withAuthList,
-  withOutAuthList,
-  FALLBACK_URL,
-  BOTTOM_MENU_INFO,
-  STATE_MODAL_TYPE_INFO,
-  USER_POINT_BTN,
-  MENU_POINT_BTN,
-  EDIT_TAB_POINT_BTN,
-  GAME_START_POINT_BTN,
-  SIGNUP_FORM,
-  SIGNUP_INITIAL_VALUES,
-  SIGNUP_VALIDATION,
-  Bounce,
-  VISIT_VALIDATION,
-  VISIT_INITIAL_VALUE,
-  LS_EMAIL,
-  LOGIN_INITIAL_VALUES,
-  LOGIN_VALIDATION,
-  LOGIN_FORM,
-  colors,
-  shadows,
-  gradients,
-}
