--- src/constants/localStorageKey.ts
@@ -1,5 +1,0 @@
-const LOCAL_STORAGE_KEY = {
-  EMAIL_CHECK_BOX: 'LS_EMAIL',
-}
-
-export default LOCAL_STORAGE_KEY
