+++ src/constants/menu/index.ts
@@ -1,0 +1,4 @@
+import { BOTTOM_MENU_INFO } from './bottom'
+import { MENU_INFO as HOME_MENU_INFO } from './home'
+
+export { BOTTOM_MENU_INFO, HOME_MENU_INFO }
