+++ src/constants/route/index.ts
@@ -1,0 +1,11 @@
+export const WITH_AUTH_URLs = [
+  '/mypage',
+  '/home',
+  '/collection',
+  '/game',
+  'island',
+  '/visit',
+]
+export const WITHOUT_AUTH_URLs = ['/signup', '/login']
+export const WITHOUT_Tabbar = ['/login', '/signup', '/', '/home']
+export const FALLBACK_URL = '/home'
