--- src/constants/routeUrl.ts
@@ -1,11 +1,0 @@
-export const withAuthList = [
-  '/mypage',
-  '/home',
-  '/collection',
-  '/game',
-  'island',
-  '/visit',
-]
-export const withOutAuthList = ['/signup', '/login']
-export const withOutTabbar = ['/login', '/signup', '/', '/home']
-export const FALLBACK_URL = '/home'
