+++ src/constants/styles/index.ts
@@ -1,0 +1,29 @@
+import { BACKGROUND_TYPE_INFO } from './background'
+import { Colors, Shadows, Gradients } from './colors'
+
+import { LOGO_TYPE_INFO } from './logo'
+import { TRIANGLE_TYPE_INFO } from './triangle'
+import { IMAGE_ADDRESS } from './imageAddress'
+import { Bounce } from './animations'
+
+import {
+  USER_POINT_BTN,
+  MENU_POINT_BTN,
+  GAME_START_POINT_BTN,
+  EDIT_TAB_POINT_BTN,
+} from './pointBtn'
+
+export {
+  IMAGE_ADDRESS,
+  BACKGROUND_TYPE_INFO,
+  LOGO_TYPE_INFO,
+  TRIANGLE_TYPE_INFO,
+  USER_POINT_BTN,
+  MENU_POINT_BTN,
+  EDIT_TAB_POINT_BTN,
+  GAME_START_POINT_BTN,
+  Bounce,
+  Colors,
+  Shadows,
+  Gradients,
+}
